Move UI loading to CoreDataAlertView
