+++ OpenGpxTracker/CoreDataHelper.swift
@@ -594,8 +594,7 @@ class CoreDataHelper {
                     alertController.addAction(cancelAction)
                     alertController.addAction(continueAction)
                     alertController.addAction(saveAction)
-                    
-                    self.showActionSheet(alertController)
+                    CoreDataAlertView().showActionSheet(alertController)
                 }
             }
             else {
@@ -603,26 +602,6 @@ class CoreDataHelper {
             }
         }
        
-    }
-    
-    /// shows the action sheet that prompts user on what to do with recovered data.
-    func showActionSheet(_ alertController: UIAlertController) {
-        let window = UIWindow(frame: UIScreen.main.bounds)
-        window.rootViewController = UIViewController()
-        
-        guard let windowLevel = UIApplication.shared.windows.last?.windowLevel else { return }
-        
-        window.windowLevel = windowLevel + 1
-        window.makeKeyAndVisible()
-        
-        if let popoverController = alertController.popoverPresentationController {
-            guard let view = window.rootViewController?.view else { return }
-            popoverController.sourceView = view
-            popoverController.sourceRect = CGRect(x: view.bounds.midX, y: view.bounds.midY, width: 0, height: 0)
-            popoverController.permittedArrowDirections = []
-        }
-        
-        window.rootViewController?.present(alertController, animated: true, completion: nil)
     }
     
     /// saves recovered data to a gpx file, silently, without loading on map.
@@ -685,6 +664,7 @@ class CoreDataHelper {
         // current segment should be 'reset' as well
         self.currentSegment = GPXTrackSegment()
         
+        // reset order sorting ids
         self.resetIds()
         
     }
